requirements.txt
# Core dependencies
click>=8.0.0
requests>=2.28.0
python-dotenv>=0.19.0
pyyaml>=6.0

# AI Integration
openai>=1.0.0

# Data processing
pandas>=1.5.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Output formatting
rich>=12.0.0
tabulate>=0.9.0

# JSON/YAML handling
pyyaml>=6.0

# Networking utilities
python-nmap>=0.7.1
dnspython>=2.2.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0